STEP2: Formalize state into CrapperKeeper
The state, which is just services and contexts, can be thought of as an
instance of the CrapperKeeper as an application.

It's a little meta, but it makes sense. Instead of tying the state of
our application to a namespace, we instead pass it around as data.

Hopefully this will allow for multiple instances of CrapperKeeper.